pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drug-prioritization-platform"
version = "0.1.0"
description = "Cloud-native multimodal AI platform for drug candidate prioritization."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Project Owner" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "PyYAML>=6.0.2",
  "pandas>=2.2.0",
  "requests>=2.32.0",
  "scikit-learn>=1.5.0",
  "rdkit>=2025.9.6",
  "mlflow>=2.10.0",
  "prometheus-fastapi-instrumentator>=7.0.0",
  "torch>=2.2.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
  "black>=24.8.0",
  "isort>=5.13.0",
  "pre-commit>=3.8.0"
]

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".venv", "mlruns"]

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
